Add Polarstern to dship command
Polarstern is introduced as additional vessel for which the shipspy dship command can be used. For this, the time stamp format was made more flexible, Polarstern was added to the ordering, additional parameters have been added to the fix function (sst_10m, dp, oxygen, turbidity), the standard deviation of heave, pitch, and roll is not mandatory anymore (since heave is missing in the Polarstern data), and 9.0 is also the missing value for Polarstern. This can be set when doing the data download.